requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.0.3
futures==3.0.5
goslate==1.5.4
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyDictionary==2.0.1
pyenchant==3.2.2
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
Werkzeug==2.2.2
